refactor: 리뷰 API 요청 Fixture 메소드 이름변경 및 유틸 클래스 추상클래스에서 final 클래스로 변경
