+++ src/test/java/com/commerce/backendserver/review/integration/ReviewApiTest.java
@@ -4,7 +4,7 @@ import static com.commerce.backendserver.common.utils.S3LinkUtils.*;
 import static com.commerce.backendserver.product.fixture.ProductFixture.*;
 import static com.commerce.backendserver.product.fixture.PromotionFixture.*;
 import static com.commerce.backendserver.review.exception.ReviewError.*;
-import static com.commerce.backendserver.review.integration.ReviewAcceptanceFixture.*;
+import static com.commerce.backendserver.review.fixture.ReviewRequestFixture.*;
 import static com.epages.restdocs.apispec.RestAssuredRestDocumentationWrapper.*;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.BDDMockito.anyString;
@@ -62,7 +62,82 @@ class ReviewApiTest extends IntegrationTestBase {
 
 		Promotion savedPromotion = promotionCommandRepository.save(VALID_FIX_PROMOTION.toEntity());
 		productRepository.save(VALID_PRODUCT.toEntity(savedPromotion));
+	}
 
+	@Nested
+	@DisplayName("[registerReview]")
+	class 리뷰_등록 {
+
+		@Test
+		@DisplayName("[Success]")
+		void success() {
+			리뷰_등록_성공(
+				spec,
+				Set.of(documentRequest()),
+				generateAccessToken(memberRepository)
+			).statusCode(CREATED.value());
+
+		}
+
+		@Test
+		@DisplayName("[Fail] 잘못된 리뷰 점수 범위로 실패")
+		void 잘못된_리뷰_점수_범위로_실패() {
+			ValidatableResponse response = 잘못된_리뷰점수_범위로_실패한다(
+				spec,
+				Set.of(documentRequest(), documentErrorResponse()),
+				generateAccessToken(memberRepository)
+			);
+
+			assertErrorResponse(response, BAD_REQUEST, INVALID_RANGE_SCORE.getMessage());
+		}
+
+		@Test
+		@DisplayName("[Fail] 잘못된 콘텐츠 길이로 실패")
+		void 잘못된_콘텐츠_길이로_실패() {
+			ValidatableResponse response = 콘텐츠_길이가_부족하여_실패한다(
+				spec,
+				Set.of(documentRequest(), documentErrorResponse()),
+				generateAccessToken(memberRepository)
+			);
+
+			assertErrorResponse(response, BAD_REQUEST, INVALID_LENGTH.getMessage());
+		}
+
+		@Test
+		@DisplayName("[Fail] 잘못된 추가정보 형식으로 실패")
+		void 잘못된_추가정보_형식으로_실패() {
+			ValidatableResponse response = 잘못된_추가정보의_형식으로_실패한다(
+				spec,
+				Set.of(documentRequest(), documentErrorResponse()),
+				generateAccessToken(memberRepository)
+			);
+
+			assertErrorResponse(response, BAD_REQUEST, INVALID_ADDITIONAL_INFO.getMessage());
+		}
+
+		@Test
+		@DisplayName("[Fail] 존재하지 않는 추가정보 이름으로 실패")
+		void 존재하지_않는_추가정보_이름으로_실패() {
+			ValidatableResponse response = 존재하지_않는_추가정보로_실패한다(
+				spec,
+				Set.of(documentRequest(), documentErrorResponse()),
+				generateAccessToken(memberRepository)
+			);
+
+			assertErrorResponse(response, BAD_REQUEST, NOT_EXIST_INFO_NAME.getMessage());
+		}
+
+		@Test
+		@DisplayName("[Fail] 숫자형 추가정보에 문자를 입력해 실패")
+		void 숫자형_추가정보에_문자를_입력해_실패() {
+			ValidatableResponse response = 숫자형_추가정보에_문자를_실패한다(
+				spec,
+				Set.of(documentRequest(), documentErrorResponse()),
+				generateAccessToken(memberRepository)
+			);
+
+			assertErrorResponse(response, BAD_REQUEST, INVALID_INTEGER_INFO_VALUE.getMessage());
+		}
 	}
 
 	private RestDocumentationFilter documentRequest() {
@@ -85,78 +160,5 @@ class ReviewApiTest extends IntegrationTestBase {
 					.attributes(constraint("지정된 이름만 사용가능"))
 			)
 		);
-	}
-
-	@Nested
-	@DisplayName("[registerReview]")
-	class 리뷰_등록 {
-
-		@Test
-		@DisplayName("[Success]")
-		void success() {
-			리뷰_등록_성공(spec, Set.of(documentRequest()), generateAccessToken(memberRepository))
-				.statusCode(CREATED.value());
-
-		}
-
-		@Test
-		@DisplayName("[Fail] 잘못된 리뷰 점수 범위로 실패")
-		void 잘못된_리뷰_점수_범위로_실패() {
-			ValidatableResponse response = failByInvalidRangeScope(
-				spec,
-				Set.of(documentRequest(), documentErrorResponse()),
-				generateAccessToken(memberRepository)
-			);
-
-			assertErrorResponse(response, BAD_REQUEST, "1에서 5사이의 정수 별점을 입력해주세요.");
-		}
-
-		@Test
-		@DisplayName("[Fail] 잘못된 콘텐츠 길이로 실패")
-		void 잘못된_콘텐츠_길이로_실패() {
-			ValidatableResponse response = failWhenPresentInvalidContentLength(
-				spec,
-				Set.of(documentRequest(), documentErrorResponse()),
-				generateAccessToken(memberRepository)
-			);
-
-			assertErrorResponse(response, BAD_REQUEST, "최소 5자 이상 입력해주세요.");
-		}
-
-		@Test
-		@DisplayName("[Fail] 잘못된 추가정보 형식으로 실패")
-		void 잘못된_추가정보_형식으로_실패() {
-			ValidatableResponse response = failWhenPresentInvalidAdditionalInfoFormat(
-				spec,
-				Set.of(documentRequest(), documentErrorResponse()),
-				generateAccessToken(memberRepository)
-			);
-
-			assertErrorResponse(response, BAD_REQUEST, INVALID_ADDITIONAL_INFO.getMessage());
-		}
-
-		@Test
-		@DisplayName("[Fail] 존재하지 않는 추가정보 이름으로 실패")
-		void 존재하지_않는_추가정보_이름으로_실패() {
-			ValidatableResponse response = failWhenPresentNonexistenceInfoName(
-				spec,
-				Set.of(documentRequest(), documentErrorResponse()),
-				generateAccessToken(memberRepository)
-			);
-
-			assertErrorResponse(response, BAD_REQUEST, NOT_EXIST_INFO_NAME.getMessage());
-		}
-
-		@Test
-		@DisplayName("[Fail] 숫자형 추가정보에 문자를 입력해 실패")
-		void 숫자형_추가정보에_문자를_입력해_실패() {
-			ValidatableResponse response = failWhenPresentInvalidIntegerInfoValue(
-				spec,
-				Set.of(documentRequest(), documentErrorResponse()),
-				generateAccessToken(memberRepository)
-			);
-
-			assertErrorResponse(response, BAD_REQUEST, INVALID_INTEGER_INFO_VALUE.getMessage());
-		}
 	}
 }
